Consolidate both viz routes
